Change ABNF for ridtxt
I think I saw some place or understood from the context, that the
relaxed rules was to sanction the current usage seen in the wild:

The current 'ALPHA / DIGIT / "." / "-"' seems overly strict, and it
does not match what's in the wild, where at least "=" "$"  "_" and
"!" has been seen by us.

I suggest to relax it further, as follows:

	ridfmt = (dot-atom-text ["@" dot-atom-text])
	       ; from RFC5322

	ridtxt = ("<" ridfmt ">") / ridfmt

This also makes it more obvious that the angle brackets are now optional.