row vs. record
The text is a bit confusing, since there is an actual "record" element
that contains a "row" element.

The text I propose to change from row(s) to record(s) refer to the record
element type that describe "row(s)" of data if you think about it in a
spreadsheet or database manner.

Possibly more clarification is needed, maybe it should be flagged for
the editor.