ABNF: domain => domain-name
Switch the definition of "receiver" and "policy-domain" from
"domain" defined in RFC 5322 to "domain-name" defined in RFC 6376.

"domain" from RFC 5322 allows for a lot of strange things as
it basically resolves to dot sepated strings of atext, which
consists of many characters not allowed in DNS labels.

   atext           =   ALPHA / DIGIT /    ; Printable US-ASCII
                       "!" / "#" /        ;  characters not including
                       "$" / "%" /        ;  specials.  Used for atoms.
                       "&" / "'" /
                       "*" / "+" /
                       "-" / "/" /
                       "=" / "?" /
                       "^" / "_" /
                       "`" / "{" /
                       "|" / "}" /
                       "~"

"domain-name" from RFC 6376, on the other hand, is much more sane.
It more concisely allows only valid DNS labels, and require at
least two labels to be combined into a valid domain.

   domain-name     = sub-domain 1*("." sub-domain)
                     ; from [RFC5321] Domain,

From RFC 5321:

   sub-domain     = Let-dig [Ldh-str]

   Let-dig        = ALPHA / DIGIT

   Ldh-str        = *( ALPHA / DIGIT / "-" ) Let-dig